Improve performance

+++ AgentConfigurationScript.ps1
@@ -1,4 +1,3 @@
-
 param([string]$agentDownloadUrl,
 	  [string]$vstsAccount,
 	  [string]$projectName,
@@ -7,53 +6,60 @@ param([string]$agentDownloadUrl,
 	  [string]$tags,
 	  [int]$agentsCount = 50)
 
-for($j=1; $j -le $agentsCount; $j++)
+$ErrorActionPreference="Stop"
+    
+If(-NOT ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] “Administrator”))
+{ 
+    throw "Run command in Administrator PowerShell Prompt"
+}
+    
+If(-NOT (Test-Path $env:SystemDrive\'vstsagent'))
 {
-    $ErrorActionPreference="Stop"
-    
-    If(-NOT ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] “Administrator”))
-    { 
-        throw "Run command in Administrator PowerShell Prompt"
-    }
-    
-    If(-NOT (Test-Path $env:SystemDrive\'vstsagent'))
-    {
-        mkdir $env:SystemDrive\'vstsagent'
-    }
-
+    mkdir $env:SystemDrive\'vstsagent'
+}
+	
+for($i=1; $i -le $agentsCount; $i++)
+{
     cd $env:SystemDrive\'vstsagent'
     $agentName = $env:COMPUTERNAME
 
-    for($i=1; $i -lt 1000; $i++)
+    for($j=$i; $j -lt 1000; $j++)
     {
-        $destFolder="A"+$i.ToString();
+        $destFolder="A"+$j.ToString();
         if(-NOT (Test-Path ($destFolder)))
         {
-            $agentName = $env:COMPUTERNAME + $i
+            $agentName = $env:COMPUTERNAME + $j
             mkdir $destFolder
             cd $destFolder
             break
         }
     }
     
-    $agentZip="$PWD\agent.zip"
+    $agentZip= $env:SystemDrive + '\vstsagent\agent.zip'
     $vstsUrl = 'https://' + $vstsAccount + '.visualstudio.com/'
-    
-    $DefaultProxy=[System.Net.WebRequest]::DefaultWebProxy;$securityProtocol=@();
-    $securityProtocol+=[Net.ServicePointManager]::SecurityProtocol;
-    $securityProtocol+=[Net.SecurityProtocolType]::Tls12;[Net.ServicePointManager]::SecurityProtocol=$securityProtocol;
-    $WebClient=New-Object Net.WebClient;
-    
-    if($DefaultProxy -and (-not $DefaultProxy.IsBypassed($agentDownloadUrl)))
-    {
-        $WebClient.Proxy= New-Object Net.WebProxy($DefaultProxy.GetProxy($agentDownloadUrl).OriginalString, $True);
-    };
 
-    $WebClient.DownloadFile($agentDownloadUrl, $agentZip);
+    if($i -eq 1)
+    {
+        # we need to download latest agent in first iteration
+        if(Test-Path $agentZip)
+        {
+            Remove-Item $agentZip;
+        }
+		
+        $DefaultProxy=[System.Net.WebRequest]::DefaultWebProxy;$securityProtocol=@();
+        $securityProtocol+=[Net.ServicePointManager]::SecurityProtocol;
+        $securityProtocol+=[Net.SecurityProtocolType]::Tls12;[Net.ServicePointManager]::SecurityProtocol=$securityProtocol;
+        $WebClient=New-Object Net.WebClient;
     
+        if($DefaultProxy -and (-not $DefaultProxy.IsBypassed($agentDownloadUrl)))
+        {
+            $WebClient.Proxy= New-Object Net.WebProxy($DefaultProxy.GetProxy($agentDownloadUrl).OriginalString, $True);
+        };
+
+        $WebClient.DownloadFile($agentDownloadUrl, $agentZip);
+    }
+	
     Add-Type -AssemblyName System.IO.Compression.FileSystem;[System.IO.Compression.ZipFile]::ExtractToDirectory( $agentZip, "$PWD");    
     
     .\config.cmd --deploymentgroup --agent $agentName --url $vstsUrl --projectname $projectName --deploymentgroupname $deploymentGroupName --auth PAT --token $PATToken --runasservice --work '_work' --unattended --adddeploymentgrouptags --deploymentgrouptags $tags
-
-    Remove-Item $agentZip;
 }
